Esempio1.jl Documentation

HelloworldJulia.Esempio1.func — Method

func(x)

Return double the number x plus 1.

e prova latec

The Schrödinger equation in a docstring:

$i\hbar\frac{\partial}{\partial t}\Psi(r,t) = \left[ \frac{-\hbar^2}{2m}\nabla^2 + V(r,t) \right] \Psi(r,t)$

This equation describes how the quantum state of a physical system changes over time.

altra prova di latec a seguire:

Maxwell's equations describe the fundamentals of electricity and magnetism. They can be written as:

\[\begin{aligned} \nabla \cdot \mathbf{E} &= 4 \pi \rho \\ \nabla \cdot \mathbf{B} &= 0 \\ \nabla \times \mathbf{E} &= - \frac{1}{c} \frac{\partial \mathbf{B}}{\partial t} \\ \nabla \times \mathbf{B} &= - \frac{1}{c} \left(4 \pi \mathbf{J} + \frac{\partial \mathbf{E}}{\partial t} \right) \end{aligned}\]

link-to-prova

prova per il link tra pagine(attraverso il titolo)

ora al contrario (attraverso una parola) funziano solo con le parole dentro un @docs, quindi fondamentalmente solo con le funzioni.